@@ -1,8 +1,8 @@
 package com.willwong.moviepages.View.viewmodel;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.annotation.NonNull;
 
 import com.willwong.moviepages.Model.Movie;
 import com.willwong.moviepages.data.database.MovieRepository;
