@@ -1,15 +1,11 @@
 package com.willwong.moviepages.View.viewmodel;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.annotation.NonNull;
 
-import com.willwong.moviepages.Model.Movie;
-import com.willwong.moviepages.Model.MoviesResponse;
 import com.willwong.moviepages.data.database.MovieRepository;
-
-import java.util.List;
 
 /**
  * Created by WillWong on 3/18/19.
