@@ -1,8 +1,8 @@
 package com.willwong.moviepages.View.viewmodel;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 
 import com.willwong.moviepages.Model.Movie;
 import com.willwong.moviepages.data.database.MovieRepository;
